Move `TracedValue` into separate module
